+++ components/org.wso2.carbon.identity.application.authenticator.samlsso/src/test/java/org/wso2/carbon/identity/application/authenticator/samlsso/artifact/SAMLSSOSoapMessageServiceTest.java
@@ -19,7 +19,6 @@
 package org.wso2.carbon.identity.application.authenticator.samlsso.artifact;
 
 import org.opensaml.Configuration;
-import org.opensaml.saml2.binding.artifact.SAML2ArtifactBuilderFactory;
 import org.opensaml.saml2.core.Artifact;
 import org.opensaml.saml2.core.ArtifactResolve;
 import org.opensaml.saml2.core.Issuer;
@@ -50,20 +49,16 @@ public class SAMLSSOSoapMessageServiceTest {
     @BeforeClass
     public void initTest() throws ArtifactResolutionException {
 
-        Map<String, String> authenticatorProperties = new HashMap<>();
-        authenticatorProperties.put(IdentityApplicationConstants.Authenticator.SAML2SSO.SP_ENTITY_ID,
-                TestConstants.SP_ENTITY_ID);
-
-        Configuration.setSAML2ArtifactBuilderFactory(new SAML2ArtifactBuilderFactory());
         Configuration.getBuilderFactory().registerBuilder(ArtifactResolve.DEFAULT_ELEMENT_NAME,
                 new ArtifactResolveBuilder());
-        Configuration.getBuilderFactory().registerBuilder(Artifact.DEFAULT_ELEMENT_NAME,
-                new ArtifactBuilder());
-        Configuration.getBuilderFactory().registerBuilder(Issuer.DEFAULT_ELEMENT_NAME,
-                new IssuerBuilder());
+        Configuration.getBuilderFactory().registerBuilder(Artifact.DEFAULT_ELEMENT_NAME, new ArtifactBuilder());
+        Configuration.getBuilderFactory().registerBuilder(Issuer.DEFAULT_ELEMENT_NAME, new IssuerBuilder());
         Configuration.getBuilderFactory().registerBuilder(Envelope.DEFAULT_ELEMENT_NAME, new EnvelopeBuilder());
         Configuration.getBuilderFactory().registerBuilder(Body.DEFAULT_ELEMENT_NAME, new BodyBuilder());
 
+        Map<String, String> authenticatorProperties = new HashMap<>();
+        authenticatorProperties.put(IdentityApplicationConstants.Authenticator.SAML2SSO.SP_ENTITY_ID,
+                TestConstants.SP_ENTITY_ID);
         SAMLSSOArtifactResolutionService artifactResolutionService = new SAMLSSOArtifactResolutionService(
                 authenticatorProperties, TestConstants.SUPER_TENANT_DOMAIN);
         artifactResolve = artifactResolutionService.generateArtifactResolveReq(TestConstants.SAML_ART);
@@ -76,5 +71,4 @@ public class SAMLSSOSoapMessageServiceTest {
         assertEquals(soapEnvelope.getBody().getUnknownXMLObjects().get(0), artifactResolve,
                 "Artifact Resolve object is not set in the soap message.");
     }
-
 }
